fix: encrypt messages

+++ src/main/java/org/cubewhy/chat/service/ChatMessageService.java
@@ -5,5 +5,5 @@ import org.cubewhy.chat.entity.ChatMessage;
 import org.cubewhy.chat.entity.dto.ChatMessageDTO;
 
 public interface ChatMessageService {
-    ChatMessage saveMessage(ChatMessageDTO message, int channelId, Account sender);
+    ChatMessage saveMessage(ChatMessageDTO message, long channelId, Account sender);
 }
